changed volume to bigint
damn you tritanium volume

+++ price db/pricefetch.py
@@ -49,7 +49,7 @@ def init():
 			`priceMax` float(16,4) DEFAULT NULL,\
 			`priceMin` float(16,4) DEFAULT NULL,\
 			`priceAverage` float(16,4) DEFAULT NULL,\
-			`volume` int(32) DEFAULT NULL,\
+			`volume` bigint(32) unsigned DEFAULT NULL,\
 			`orders` int(16) DEFAULT NULL,\
 			`priceOpen` float(16,4) DEFAULT NULL,\
 			`priceClose` float(16,4) DEFAULT NULL,\
